backend/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bokeh==3.1.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clarifai==9.8.0
clarifai-grpc==9.8.0
click==8.1.3
cloudpickle==2.2.1
cohere==4.24
comm==0.1.3
contourpy==1.0.7
cryptography==42.0.5
dask==2023.4.1
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distributed==2023.4.1
dnspython==2.4.2
duckdb==0.7.1
executing==1.2.0
falcon==1.4.1
falcon-auth==1.1.0
falcon-autocrud==1.0.37
falcon-cors==1.1.7
falcon-limiter==1.0.1
falcon-openapi==0.5.1
fastavro==1.8.2
fastjsonschema==2.16.3
filelock==3.12.3
Flask==2.3.2
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.5.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.58.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.6.0
importlib_resources==6.1.2
install==1.3.5
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
jwt==1.3.1
langchain==0.0.285
langsmith==0.0.35
limits==3.9.0
locket==1.0.0
loguru==0.7.1
lz4==4.3.2
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nlpcloud==1.1.44
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.24.3
openai==0.28.0
openlm==0.0.5
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.5.0
polars==0.17.12
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.15.1
PyJWT==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-mimeparse==1.6.0
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
redis==5.0.0
regex==2023.8.8
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
safetensors==0.3.3
Send2Trash==1.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlitedict==2.1.0
stack-data==0.6.2
sympy==1.12
tblib==1.7.0
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.33.1
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.16.0
xxhash==3.3.0
xyzservices==2023.2.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zict==3.0.0
zipp==3.15.0
falcon_apispec
gunicorn